Face rotations now in rotations.py
All the algorithms pertaining to rotating the faces of the cube are now in rotations.py rather than cluttering main.py
